apps/app_backend/requirements.txt
Django==4.2.11
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.1.0
pandas==2.2
scikit-learn==1.4.2
gunicorn==21.0.1
pytest-django==4.8.0